+++ src/view/containers/GameModal/index.js
