+++ src/view/containers/GameModal/modalEvent.js
